make userclick and script URL parameters, pass the rest as form data
By passing a URLSearchParams object to fetch() directly instead
of converting it to a string, it is encoded as form data and
NS accepts it properly. 'chk' and 'localid' seem to be needed
in the form data (and not as URL parameters), while the OSRS
indicate that 'script' and 'userclick' should be URL
search parameters.