Fix the NS HTML request engine and support fast site (#3)
* make userclick and script URL parameters, pass the rest as form data

By passing a URLSearchParams object to fetch() directly instead
of converting it to a string, it is encoded as form data and
NS accepts it properly. 'chk' and 'localid' seem to be needed
in the form data (and not as URL parameters), while the OSRS
indicate that 'script' and 'userclick' should be URL
search parameters.

* make NS HTML requests compatible with the fast site

* show endo success properly

fetch with redirects set to 'manual' will NOT
include the URL you need to navigate to, unfortunately,
so the only way to check success is to follow the
redirect and check the response's URL field